Adding fatigue tests
