+++ pyFAST/postpro/__init__.py
@@ -1,2 +1,3 @@
 # For now making everything available
 from pyFAST.postpro.postpro import *
+from pyFAST.tools.fatigue import equivalent_load
